LayoutTests/http/tests/inspector/network/network-disable-cache-xhrs.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function loadXHR(callback)
{
    doXHR("GET", "resources/resource.php?random=1&cached=1", true, callback);
}

function loadData()
{
    loadXHR(xhrLoaded1);
}

function xhrLoaded1()
{
    loadXHR(xhrLoaded2);
}

function xhrLoaded2()
{
    console.log("First XHR loaded.");
}

function cacheDisabled()
{
    loadXHR(allXHRsLoaded);
}

function allXHRsLoaded()
{
    console.log("Done.");
}

function test()
{
    // Since this test could be run together with other inspector backend cache
    // tests, we need to reset size limits to default ones.
    InspectorTest.resetInspectorResourcesData(step1);

    function step1()
    {
        InspectorTest.addConsoleSniffer(step2);
        InspectorTest.evaluateInPage("loadData()");
    }

    function step2(msg)
    {
        NetworkAgent.setCacheDisabled(true, step3);
    }

    function step3(msg)
    {
        InspectorTest.addConsoleSniffer(step4);
        InspectorTest.evaluateInPage("cacheDisabled()");
    }

    function step4(msg)
    {
        // inspector-test.js appears in network panel occasionally in Safari on
        // Mac, so checking two last resources.
        var requestsCount = WebInspector.panel("network").requests.length;
        var request1 = WebInspector.panel("network").requests[requestsCount - 3];
        var request2 = WebInspector.panel("network").requests[requestsCount - 2];
        var request3 = WebInspector.panel("network").requests[requestsCount - 1];
        request1.requestContent(contentLoaded);
        request2.requestContent(contentLoaded);
        request3.requestContent(contentLoaded);
    }

    var contentLoadedCount = 0;
    function contentLoaded()
    {
        if (++contentLoadedCount !== 3)
            return;

        var requestsCount = WebInspector.panel("network").requests.length;
        var request1 = WebInspector.panel("network").requests[requestsCount - 3];
        var request2 = WebInspector.panel("network").requests[requestsCount - 2];
        var request3 = WebInspector.panel("network").requests[requestsCount - 1];

        InspectorTest.addResult(request1.url);
        InspectorTest.addResult(request2.url);
        InspectorTest.addResult(request3.url);
        InspectorTest.assertTrue(request1.content === request2.content, "First and second resources are equal");
        InspectorTest.assertTrue(request2.content !== request3.content, "Second and third resources differ");
        NetworkAgent.setCacheDisabled(false, step5);
    }

    function step5(msg)
    {
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
    <p>Tests disabling cache from inspector.</p>
</body>
</html>
